requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
aitextgen==0.5.2
anyio==3.4.0
asgiref==3.4.1
astroid==2.9.0
async-timeout==4.0.2
attrs==21.2.0
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
fastapi==0.70.1
filelock==3.4.2
fire==0.4.0
frozenlist==1.2.0
fsspec==2021.11.1
future==0.18.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.43.0
h11==0.12.0
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.10.0
isort==5.10.1
joblib==1.1.0
lazy-object-proxy==1.7.1
Markdown==3.3.6
mccabe==0.6.1
multidict==5.2.0
mypy-extensions==0.4.3
numpy==1.21.5
oauthlib==3.1.1
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyDeprecate==0.3.1
pylint==2.12.2
pyparsing==3.0.6
python-dotenv==0.19.2
pytorch-lightning==1.5.7
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
tokenizers==0.10.3
toml==0.10.2
tomli==1.2.3
torch==1.10.1
torchmetrics==0.6.2
tqdm==4.62.3
transformers==4.15.0
treelib==1.6.1
typing-extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
Werkzeug==2.0.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
